+++ api/RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 ## in develop
 
+* Handling of non-fully qualified file IDs for bulk search/describe. Now for the files provided without the project ID, 
+the `describe` response will be returned only for current workspace/project. If a file was cloned to other projects, they 
+will be ignored. Non-fully qualified file IDs are not allowed when searching files in other projects.
+
 * Regression tests for API calls to platform